Custom_Script/requirements.txt
azureml-sdk>=1.10.0
azureml-opendatasets>=1.10.0
azureml-contrib-pipeline-steps>=1.10.0
jupyter
matplotlib
numpy
pandas
seaborn
